Groups tree optimization
* The tree view of groups was not optimized for a large number of
groups. With the number of groups being high (e.g. 1000) the whole
application would lag and freeze.
* The tree view component has been reworked so it uses a virtual scoll.
This improves the performance by a lot. I have tested this changes on a
vo with 3000 groups and it works much better.
* I have also removed the check that the group description is shown only
if it doesn't fit. This check caused some errors because it used native
html properties to detect this. Also, this way the behaviour is more
consistent.