Fix applications list for many applications
* The paginator on the application list was implemented so it was not
actually used for the first view. The paginator was undefined during the
time it was set to the datasource. This caused the first view to render
all of the applications.
* On vo applications and group applications I have also replaces the
filter by a debounce filter, which is better to use with such large
number of applications. The filtration takes a long time by default.